I think I fixed the simulation
So I think I found and fixed the problem. The problem was I flipped the variables when defining the escape length in such a way that it was returning a negative integer. Since the code automatically sets the length to the escape length if the length is bigger than the escape length it caused a lot of the length values to be negative making the spark number negative. Now the simulation appears to only generate spark numbers between 0 and 7 for different values of Muon mass.